:recycle:refactor : winner를 전략이 아닌 객체로 수정
